# openapi.yaml
openapi: 3.0.3
info:
  title: AquaVerse AI API Contract
  version: 1.0.0
  description: Executable API contract for AquaVerse AI analyst dashboard, mobile client, digital twin, and M1/M2/M3 reasoning adapters.
paths:
  /v1/auth/token:
    post:
      summary: Staff login
      operationId: loginToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Keycloak role-scoped JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string }
                  role: { type: string }
  /v1/auth/otp/request:
    post:
      summary: Request OTP for farmer phone login
      operationId: requestOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        '200':
          description: OTP request sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  txn_id: { type: string }
  /v1/auth/otp/verify:
    post:
      summary: Verify OTP code and issue JWT token
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, otp, txn_id]
              properties:
                phone: { type: string }
                otp: { type: string }
                txn_id: { type: string }
      responses:
        '200':
          description: Verified farmer JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string }
                  role: { type: string }
                  farmer_name: { type: string }
  /v1/ponds:
    get:
      summary: Paginated list of ponds
      operationId: listPonds
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: district
          in: query
          schema: { type: string }
        - name: taluk
          in: query
          schema: { type: string }
        - name: species
          in: query
          schema: { type: string }
        - name: doc_band
          in: query
          schema: { type: string }
        - name: risk_band
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated pond list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PondRecord' }
                  next_cursor: { type: string, nullable: true }
  /v1/ponds/{pond_id}:
    get:
      summary: Full pond detail
      operationId: getPondDetail
      parameters:
        - name: pond_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full pond record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PondRecord' }
  /v1/ponds/{pond_id}/timeseries:
    get:
      summary: Multi-parameter timeseries
      operationId: getPondTimeseries
      parameters:
        - name: pond_id
          in: path
          required: true
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string }
        - name: to
          in: query
          schema: { type: string }
        - name: params
          in: query
          schema: { type: string }
        - name: agg
          in: query
          schema: { type: string, enum: [raw, hourly, daily] }
      responses:
        '200':
          description: Multi-parameter series payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  timestamps: { type: array, items: { type: string } }
                  series:
                    type: object
                    additionalProperties:
                      type: array
                      items: { type: number, nullable: true }
...
  /v1/logs:
    post:
      summary: Water-quality telemetry & manual log ingestion
      operationId: ingestLog
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WaterQualityLog' }
      responses:
        '201':
          description: Log ingested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  log_id: { type: string }
                  status: { type: string }
                  timestamp: { type: string }
  /v1/media/presign:
    post:
      summary: Generate MinIO / R2 presigned URL for pond imagery & sensor uploads
      operationId: presignMediaUpload
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignRequest' }
      responses:
        '200':
          description: Presigned upload URL generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PresignResponse' }
  /v1/media/commit:
    post:
      summary: Commit uploaded media asset to pond metadata
      operationId: commitMediaUpload
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaCommit' }
      responses:
        '200':
          description: Media committed
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id: { type: string }
                  status: { type: string }
...
  /v1/forecast/temporal:
    get:
      summary: Multi-variable temporal forecast (TFT / TCN / PatchTST)
      operationId: getTemporalForecast
      parameters:
        - name: pond_id
          in: query
          required: true
          schema: { type: string }
        - name: model_family
          in: query
          schema: { type: string, enum: [TFT, TCN, PatchTST], default: TFT }
        - name: horizon_hours
          in: query
          schema: { type: integer, default: 72 }
      responses:
        '200':
          description: Temporal multi-variable forecast payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TemporalForecastPayload' }
...
  /v1/twin/{pond_id}/state:
    get:
      summary: Digital twin state vector at time t
      operationId: getTwinState
      parameters:
        - name: pond_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Twin state vector
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwinState' }
  /v1/twin/{pond_id}/whatif:
    post:
      summary: Re-run simulator server-side under intervention
      operationId: runTwinWhatIf
      parameters:
        - name: pond_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                water_exchange_rate: { type: number }
                aeration_hours: { type: number }
                feed_reduction_pct: { type: number }
      responses:
        '200':
          description: Updated state series under intervention
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwinState' }
  /v1/reason:
    post:
      summary: Internal reasoning call using hot-swapped LoRA adapters
      operationId: getReasoning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pond_id, adapter]
              properties:
                pond_id: { type: string }
                adapter:
                  type: string
                  enum: [m1_chemistry, m2_health, m3_production]
      responses:
        '200':
          description: M1/M2/M3 reasoning response
          content:
            application/json:
              schema:
                type: object
                properties:
                  explanation: { type: string }
                  reasoning_trace: { type: string }
                  recommendations: { type: array, items: { type: string } }
  /v1/ask:
    post:
      summary: Farmer-facing conversational Q&A endpoint powered by vLLM (Qwen3-8B)
      operationId: askFarmerAssistant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AskQuery' }
      responses:
        '200':
          description: Farmer assistant response with optional Indic translation & audio TTS
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AskResponse' }
...
  /v1/translate:
    post:
      summary: IndicTrans2 translation service (English -> Tamil/Telugu/Hindi)
      operationId: translateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, target_lang]
              properties:
                text: { type: string }
                target_lang: { type: string, enum: [ta, te, hi] }
      responses:
        '200':
          description: Translated output with cache status
          content:
            application/json:
              schema:
                type: object
                properties:
                  translated_text: { type: string }
                  cached: { type: boolean }
  /v1/reports/export:
    get:
      summary: Export PDF or XLSX in official government letterhead format
      operationId: exportReport
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [pdf, xlsx], default: pdf }
        - name: district
          in: query
          schema: { type: string }
      responses:
        '200':
          description: File download trigger
          content:
            application/json:
              schema: { type: object, properties: { download_url: { type: string } } }
components:
  schemas:
    PondRecord:
      type: object
      properties:
        pond_id: { type: string }
        farmer_name: { type: string }
        district: { type: string }
        taluk: { type: string }
        species: { type: string }
        doc: { type: integer }
        area_ha: { type: number }
        depth_m: { type: number }
        water_source: { type: string }
        biomass_kg: { type: number }
        current_risk: { type: number }
        risk_band: { type: string }
...
    TwinState:
      type: object
      properties:
        pond_id: { type: string }
        timestamp: { type: string }
        do_mg_l: { type: number }
        ph: { type: number }
        temp_c: { type: number }
        tan_mg_l: { type: number }
        no2_mg_l: { type: number }
        salinity_ppt: { type: number }
        alkalinity_mg_l: { type: number }
        biomass_kg: { type: number }
        water_clarity_cm: { type: number }
...
    WaterQualityLog:
      type: object
      required: [pond_id, do_mg_l, ph, temp_c]
      properties:
        pond_id: { type: string }
        timestamp: { type: string }
        do_mg_l: { type: number }
        ph: { type: number }
        temp_c: { type: number }
        tan_mg_l: { type: number }
        salinity_ppt: { type: number }
        notes: { type: string }
        source: { type: string, enum: [manual, iot_sensor, drone_flyover] }
    PresignRequest:
      type: object
      required: [pond_id, filename, content_type]
      properties:
        pond_id: { type: string }
        filename: { type: string }
        content_type: { type: string }
        category: { type: string, enum: [water_sample, microscopic_slide, pond_perimeter, disease_observation] }
    PresignResponse:
      type: object
      properties:
        upload_url: { type: string }
        asset_key: { type: string }
        expires_at: { type: string }
    MediaCommit:
      type: object
      required: [pond_id, asset_key]
      properties:
        pond_id: { type: string }
        asset_key: { type: string }
        caption: { type: string }
    TemporalForecastPayload:
      type: object
      properties:
        pond_id: { type: string }
        model_family: { type: string }
        horizon_hours: { type: integer }
        timestamps: { type: array, items: { type: string } }
        forecast_do: { type: array, items: { type: number } }
        forecast_ph: { type: array, items: { type: number } }
        forecast_temp: { type: array, items: { type: number } }
        confidence_interval:
          type: object
          properties:
            do_p10: { type: array, items: { type: number } }
            do_p90: { type: array, items: { type: number } }
    AskQuery:
      type: object
      required: [question]
      properties:
        question: { type: string }
        pond_id: { type: string }
        lang: { type: string, enum: [en, ta, te, hi], default: en }
        voice_tts: { type: boolean, default: false }
    AskResponse:
      type: object
      properties:
        answer: { type: string }
        translated_answer: { type: string }
        audio_url: { type: string, nullable: true }
        confidence: { type: number }
        sources: { type: array, items: { type: string } }